Add combined reducers
